+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.controller.dto.CreateReservationTimeRequest;
+import roomescape.dto.CreateReservationTimeRequest;
 import roomescape.domain.ReservationTime;
 import roomescape.service.ReservationTimeService;
