+++ src/main/java/roomescape/dto/CreateReservationRequest.java
@@ -1,4 +1,4 @@
-package roomescape.controller.dto;
+package roomescape.dto;
 
 public record CreateReservationRequest(String date, String name, Long timeId) {
