+++ src/main/java/roomescape/repository/ReservationRepository.java
@@ -6,7 +6,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
-import roomescape.controller.dto.CreateReservationRequest;
+import roomescape.dto.CreateReservationRequest;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
